requirements.txt
pysrt
translators
pywebview
torch
whisper
stable-ts
webvtt-py